First seemingly working version
This version works (or at least it compiles).

Potentially not an ideal implementation as I've had to create separate
versions of the ApplyToMVN function.

A better implementation might be to add laplacian weighting to the
context after all (I initially tried this and decided against it). This
will have the benefit of not needing to modify ApplyToMVN or CalculateaK
to accept more arguments.